+++ cmd/rescale-int/main.go
@@ -14,7 +14,7 @@ import (
 
 // Version information
 var (
-	Version   = "v4.3.7"
+	Version   = "v4.3.8"
 	BuildTime = "2026-01-14"
 )